Add more structure and move common tools out of mysql
