+++ src/unix/command-emulation.test.ts
